% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yasegfit.R
\name{fitted.yasegfit}
\alias{fitted.yasegfit}
\title{Extract the fitted values from a yasegfit object}
\usage{
\method{fitted}{yasegfit}(sf, concat = TRUE)
}
\arguments{
\item{sf}{The yasegfit object}

\item{concat}{Whether the fitted values of different segments will be
concatenated.}
}
\value{
If concat==TRUE, the fitted values of different segments
              will be concatenated into one vector. Otherwise,
              the returned value will be a list of fitted values
              of each segment.
}
\description{
Extract the fitted values from a yasegfit object
}
\details{
Extract the fitted values from a yasegfit object
}
