% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yasegfit.R
\name{plot.yasegfit}
\alias{plot.yasegfit}
\title{Plot a segfit object}
\usage{
\method{plot}{yasegfit}(x, y = "", col.data = "black", col.seg = "red",
  legend.pos = "topleft")
}
\arguments{
\item{x}{If x is a segfit object, input y is ignored.}

\item{y}{A yasegfit object, If not provided, x must
be a yasegfit object.}

\item{col.data}{color of data}

\item{col.seg}{color of segments}

\item{legend.pos}{legend positions}
}
\description{
Plot a yasegfit object
}
\details{
Plot a yasegfit object
}
