% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yasegfit.R
\name{residuals.yasegfit}
\alias{residuals.yasegfit}
\title{Extract the residuals from a yasegfit object}
\usage{
\method{residuals}{yasegfit}(sf, concat = TRUE)
}
\arguments{
\item{sf}{The yasegfit object}

\item{concat}{Whether the residuals of different segments will be
concatenated.}
}
\value{
If concat==TRUE, the residuals of different segments
              will be concatenated into one vector. Otherwise,
              the returned value will be a list of residuals
              of each segment.
}
\description{
Extract the residuals from a yasegfit object
}
\details{
Extract the residuals from a yasegfit object
}
