% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yasegfit.R
\name{summary.yasegfit}
\alias{summary.yasegfit}
\title{Summarise a yasegfit object}
\usage{
\method{summary}{yasegfit}(sf)
}
\arguments{
\item{sf}{The yasegfit object to summarise}
}
\description{
Summarise a yasegfit object
}
\details{
Summarise a yasegfit object
}
