% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yasegfit.R
\name{yasegfit}
\alias{yasegfit}
\title{Segfit a series}
\usage{
yasegfit(data, seg.alg = "dp", fit.alg = "abline", smp = 2.3, nseg = 3)
}
\arguments{
\item{data}{The input series to be segfitted.}

\item{seg.alg}{Segmentation algorithm, possible values: c("topdown", "dp")}

\item{fit.alg}{Fitting algorithm for each segment,
possible values: c("abline", "powerabc", "const")}

\item{smp}{Controlling the number of segments when seg.alg="fit"}

\item{nseg}{The exact number of segments when seg.alg="topdown"}
}
\value{
A yasegfit object containing various kinds of information.
        For each segment, it contains the head index (hi), tail index (ei),
        parameter $a$ (a) parameter $b$ (b), parameter $c$ (c), fitting order
        (order), fitted values (fit), and fitting residuals (residual).
        It also has a attribute "data" for the original data; A attribute
        "seg.alg" for dp algorithm used; A attribute "fit.alg" for fitting
        algorithm used for each segment; A attribute "smp" for the smp used
        in "dp" algorithm; A attribute "nseg" for number of segments.
}
\description{
Segfit a series
}
\details{
Segfit a series
}
